backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
anthropic>=0.40.0
langdetect==1.0.9
python-dotenv==1.0.1
python-pptx==1.0.0
pydantic==2.7.0
pypdf==4.3.1
pymupdf==1.24.5
python-docx==1.1.2
python-multipart==0.0.9
